refactor(api): Move setup_required import to wraps module across multiple files
